Change `ClientConfigFactory` to an interface implemented by `DefaultClientConfigFactory`
Makes it possible to modify settings not governed by properties, e.g. by using decorator pattern.
